+++ packages/website/src/examples/MultiSelect/withFilter.example.tsx
@@ -1,0 +1,14 @@
+import {MultiSelectWithFilter} from '@coveord/plasma-react';
+import * as React from 'react';
+
+export default () => (
+    <MultiSelectWithFilter
+        id="mutli-select-2"
+        customValues
+        items={[
+            {value: 'one', displayValue: 'Option 1'},
+            {value: 'two', displayValue: 'Option 2'},
+            {value: 'three', displayValue: 'Option 3'},
+        ]}
+    />
+);