+++ packages/website/src/examples/Slider/Slider.example.tsx
@@ -1,0 +1,17 @@
+import {Slider} from '@coveord/plasma-react';
+import * as React from 'react';
+
+export default () => (
+    <Slider
+        id="🍄"
+        hasTooltip
+        step={25}
+        marks={{
+            0: 'Lower',
+            50: 'Middle',
+            100: 'Higher',
+        }}
+        initialValue={50}
+        appendValueFormatter={(value: number) => `${value}$`}
+    />
+);