+++ packages/website/src/pages/feedback/LastUpdated.tsx
@@ -1,34 +1,16 @@
+import code from '@examples/LastUpdated/LastUpdated.example.tsx';
+import specificTime from '@examples/LastUpdated/LastUpdatedSpecifiedTime.example.tsx';
 import * as React from 'react';
-import {LastUpdated} from '@coveord/plasma-react';
+import {PageLayout} from '../../building-blocs/PageLayout';
 
-import PlasmaComponent from '../../building-blocs/PlasmaComponent';
-
-// start-print
-
-export const LastUpdatedExamples: React.FunctionComponent = () => (
-    <PlasmaComponent
+export default () => (
+    <PageLayout
         id="LastUpdated"
         title="Last Updated"
-        usage="A “last updated” string displays the time a set of data has been last updated by a system."
-        withSource
-    >
-        <div className="mt2" style={{width: 400}}>
-            <div className="form-group">
-                <label className="form-control-label">Last update</label>
-                <LastUpdated id="LastUpdatedExampleComponent" />
-            </div>
-            <div className="form-group">
-                <label className="form-control-label">Last update with custom time</label>
-                <LastUpdated
-                    id="LastUpdatedExampleComponentWithTime"
-                    time={new Date(+new Date() - Math.floor(Math.random() * 10000000000))}
-                />
-            </div>
-            <div className="form-group">
-                <label className="form-control-label">Last update with label changed</label>
-                <LastUpdated id="LastUpdatedExampleComponentWithLabel" label="Dernière modification à" />
-            </div>
-        </div>
-    </PlasmaComponent>
+        section="Feedback"
+        description="A “last updated” string displays the time a set of data has been last updated by a system."
+        componentSourcePath="/lastUpdated/LastUpdated.tsx"
+        code={code}
+        examples={{specificTime: {code: specificTime, title: 'Specific date'}}}
+    />
 );
-export default LastUpdatedExamples;