+++ packages/website/src/pages/form/JSONEditor.tsx
@@ -1,39 +1,23 @@
+import code from '@examples/JSONEditor/JSONEditor.example.tsx';
+import readOnly from '@examples/JSONEditor/ReadOnly.example.tsx';
+import inError from '@examples/JSONEditor/InError.example.tsx';
+import valueFromState from '@examples/JSONEditor/ValueState.example.tsx';
 import * as React from 'react';
-import {JSONEditor} from '@coveord/plasma-react';
 
-import PlasmaComponent from '../../building-blocs/PlasmaComponent';
-import {fakeJSON, JSONToString} from '../../utils/DiffViewerExmaplesCommon';
+import {PageLayout} from '../../building-blocs/PageLayout';
 
-// start-print
-export class JSONEditorExamples extends React.Component {
-    render() {
-        return (
-            <PlasmaComponent
-                id="JSONEditor"
-                title="JSON Editor"
-                usage="A JSON editor is a text area where users can enter and edit data in JSON format."
-                withSource
-            >
-                <div className="mt2">
-                    <div className="form-group">
-                        <label className="form-control-label">JSON Editor using codemirror</label>
-                        <JSONEditor value={JSONToString(fakeJSON)} />
-                    </div>
-
-                    <div className="form-group">
-                        <label className="form-control-label">JSON Editor using codemirror in readonly mode</label>
-                        <JSONEditor value={JSONToString(fakeJSON)} readOnly />
-                    </div>
-
-                    <div className="form-group">
-                        <label className="form-control-label">
-                            JSON Editor using codemirror with an action on change
-                        </label>
-                        <JSONEditor value={JSONToString(fakeJSON)} onChange={(json: string) => alert(json)} />
-                    </div>
-                </div>
-            </PlasmaComponent>
-        );
-    }
-}
-export default JSONEditorExamples;
+export default () => (
+    <PageLayout
+        id="JSONEditorConnected"
+        title="JSON Editor"
+        section="Form"
+        description="A JSON editor is a text area where users can enter and edit data in JSON format."
+        componentSourcePath="/editor/JSONEditor.tsx"
+        code={code}
+        examples={{
+            readOnly: {code: readOnly, title: 'Read only'},
+            inError: {code: inError, title: 'Error Message'},
+            valueFromState: {code: valueFromState, title: 'Selector'},
+        }}
+    />
+);